A2_ReAct/requirements.txt
tavily-python>=0.3.3  # or higher
Groq
dotenv